revise structure into MVC
