+++ app/main.py
@@ -1,13 +1,13 @@
 from fastapi import *
 from fastapi.responses import FileResponse
 from fastapi.staticfiles import StaticFiles
-from app.routes import attractionID, attractions, mrts, userAuth
+
+from app.routes import attractions, metros, userAuth
 
 app=FastAPI()
 
 app.include_router(attractions.router)
-app.include_router(attractionID.router)
-app.include_router(mrts.router)
+app.include_router(metros.router)
 app.include_router(userAuth.router)
 
 # Static Pages (Never Modify Code in this Block)
