+++ app/routes/userAuth.py
@@ -4,11 +4,14 @@ import mysql.connector
 from fastapi.security import OAuth2PasswordBearer
 
 from app.database import get_db_connection
+from app.model.auth import Auth
+from app.model.auth_token import AuthToken
+
 
 router = APIRouter()
 
 @router.post("/api/user")
-async def sign_out(request:Request):
+async def sign_up(request:Request):
     body = await request.json()
     # print(body)
     name = body.get('name')
@@ -21,12 +24,9 @@ async def sign_out(request:Request):
             status_code=400
         )
     try:
-        with get_db_connection() as db:
-            with db.cursor(dictionary=True) as cursor:
-                cursor.execute("INSERT INTO userData (userName, userMail, userPassword) VALUES (%s, %s, %s);",(name, mail, password))
-                db.commit()       
-                print(f"註冊成功 {mail} 已加入資料庫")
-                return{ "ok": True }
+        signup_is_success = Auth.user_sign_up(name, mail, password)
+        if signup_is_success:
+            return{ "ok": True }
     except mysql.connector.IntegrityError as e:
         print(f"註冊失敗")
         return JSONResponse(
@@ -47,8 +47,6 @@ async def sign_out(request:Request):
 
 @router.put("/api/user/auth")
 async def sign_in(request:Request):
-    from app.login_token import create_jwt_token
-
     body = await request.json(); 
     # print(body)
     mail = body.get('email')
@@ -59,19 +57,15 @@ async def sign_in(request:Request):
             status_code=400
         )
     try:
-        with get_db_connection() as db:
-            with db.cursor(dictionary=True) as cursor:
-                cursor.execute("SELECT * FROM userData WHERE userMail=%s AND userPassword=%s;",(mail, password))
-                userData = cursor.fetchone()
-                if not userData:
-                    print("登入失敗")
-                    return JSONResponse(content={ "error": True, "message": "登入失敗，帳號或密碼錯誤"}, status_code=400)     
-                token = create_jwt_token({
-                    "mail":userData['userMail'],
-                    "name":userData['userName'],
-                    "id":userData['id']
-                })
-        # print("token: ", token)
+        success_signin = Auth.user_sign_in(mail, password)
+        if not success_signin:
+            print("登入失敗")
+            return JSONResponse(content={ "error": True, "message": "登入失敗，帳號或密碼錯誤"}, status_code=400)     
+        token = AuthToken.create_jwt_token({
+            "mail":success_signin['userMail'],
+            "name":success_signin['userName'],
+            "id":success_signin['id']
+        })
         return {"token":token}
     except Exception as e:
         print(f"Error: {e}")
@@ -88,9 +82,8 @@ oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 
 @router.get("/api/user/auth")
 async def get_user_auth(token:str = Depends(oauth2_scheme)):
-    from app.login_token import verify_jwt_token
     try:
-        user_data = verify_jwt_token(token)
+        user_data = AuthToken.verify_jwt_token(token)
         data ={
             "id":user_data['id'],
             "name":user_data['name'],
